8047130: Fewer escapes from escape analysis
Treat max_stack attribute as an int in bytecode escape analyzer

+++ hotspot/src/share/vm/ci/bcEscapeAnalyzer.cpp
@@ -89,8 +89,8 @@ class BCEscapeAnalyzer::StateInfo {
 public:
   ArgumentMap *_vars;
   ArgumentMap *_stack;
-  short _stack_height;
-  short _max_stack;
+  int _stack_height;
+  int _max_stack;
   bool _initialized;
   ArgumentMap empty_map;
